Added auto refresh to the activity
OnResume creates an AsyncTask which keeps running
until cancelled. It requests server updates with
a delay of two seconds between each update.

OnPause cancels the AsyncTask.

To handle multiselect gracefully, the fragment requests
the auto refresh to be paused. This pause is remembered
between resets, as long as the multiselect mode is active.
Once the users selects an action or exits multiselect,
the auto refresh resumes.

There are probably several things I haven't thought of,
and there should be a user setting for this maybe.

Signed-off-by: Jonas Kalderstam <jonas@kalderstam.se>